requirements.txt
python-telegram-bot==13.7
pandas==1.2.4
seaborn==0.11.1
matplotlib==3.3.4
pdfkit==0.6.1